prediction/models/ImageShapeMicroservice/requirements.txt
pydantic
fastapi
pyinstaller
setuptools
uuid
uvicorn
python-multipart
pillow
python-dotenv
requests
starlette
rq
redis